requirements.txt
streamlit
Pillow
numpy
tensorflow
matplotlib
seaborn
pandas